+++ pczt/src/transparent.rs
@@ -41,6 +41,7 @@ pub struct Input {
     /// - This is set by the Constructor.
     /// - If omitted, the sequence number is assumed to be the final sequence number
     ///   (`0xffffffff`).
+    #[getset(get = "pub")]
     pub(crate) sequence: Option<u32>,
 
     /// The minimum Unix timstamp that this input requires to be set as the transaction's
@@ -66,6 +67,7 @@ pub struct Input {
     // needed for computing the binding signatures.
     #[getset(get = "pub")]
     pub(crate) value: u64,
+    #[getset(get = "pub")]
     pub(crate) script_pubkey: Vec<u8>,
 
     /// The script required to spend this output, if it is P2SH.
@@ -139,7 +141,9 @@ pub struct Output {
     // These are required fields that are part of the final transaction, and are filled in
     // by the Constructor when adding an output.
     //
+    #[getset(get = "pub")]
     pub(crate) value: u64,
+    #[getset(get = "pub")]
     pub(crate) script_pubkey: Vec<u8>,
 
     /// The script required to spend this output, if it is P2SH.