+++ zcash_client_backend/src/scanning.rs
@@ -1187,7 +1187,7 @@ pub mod testing {
         };
         let fake_cmu = {
             let fake_cmu = bls12_381::Scalar::random(&mut rng);
-            fake_cmu.to_repr().as_ref().to_owned()
+            fake_cmu.to_repr().to_vec()
         };
         let fake_epk = {
             let mut buffer = [0; 64];
@@ -1266,7 +1266,7 @@ pub mod testing {
         let cout = CompactSaplingOutput {
             cmu,
             ephemeral_key,
-            ciphertext: enc_ciphertext.as_ref()[..52].to_vec(),
+            ciphertext: enc_ciphertext[..52].to_vec(),
         };
         let mut ctx = CompactTx::default();
         let mut txid = vec![0; 32];