+++ zcash_history/CHANGELOG.md
@@ -7,7 +7,7 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 ### Changed
-- MSRV is now 1.77.0.
+- MSRV is now 1.81.0.
 
 ## [0.4.0] - 2023-03-01
 ### Changed